backend/requirements.txt
fastapi~=0.92.0
uvicorn~=0.20.0
tortoise-orm[asyncpg]~=0.19.3